Refactor: moved queryDataTable inside dbconnector
